cpufreq: cpu-boost: Integrate Dynamic SchedTune
This commit introduces Dynamic SchedTune boosting to cpu-boost driver that has been deprecated due to the removal of HMP. The decision for choosing this driver integration is due to the support of various kernel managers to this driver's module_params as well as the correlation of the concept of this idea to what this driver's purpose.
Integrating Dynamic SchedTune is advantageous in a way that we could adjust the amount of boost that is to be used during runtime depending on event, e.g. a touch input.

This driver has the following features:
- Separated boost triggers corresponding to specific events.
- Detection of task zygotes by Sultan and boosting each time such is detected in top-app cgroup.
- Adjustable sched_boost and duration variables.
- Disable_dsboost boolean for disabling the functions of the driver.
- Serialized work_structs execution.

Signed-off-by: ederekun <sedrickvince@gmail.com>